requirements.txt

pandas>=2.2
numpy>=1.26
matplotlib>=3.8
scipy>=1.11
statsmodels>=0.14
beautifulsoup4>=4.12
feedparser>=6.0
requests>=2.32
python-dateutil>=2.9
pytz>=2024.1
PyYAML>=6.0
fredapi>=0.5
yfinance>=0.2
jinja2>=3.1
markdown>=3.6
